Changed to floats

+++ src/DistributedCoordinationFunction.c
@@ -1,8 +1,8 @@
 #include "DistributedCoordinationFunction.h"
 
-int * generatePoissonDelayTimes(int lambda, int time)
+float * generatePoissonDelayTimes(int lambda, int time)
 {
-	int * delayTimes = (int *) calloc(sizeof(int), lambda*time);
+	float * delayTimes = (float *) calloc(sizeof(float), lambda*time);
 
 	for(int i = 0; i < lambda*time; i++) {
 		delayTimes[i] = -1.0 / lambda * log(1 - (rand()%100 / 100.0));
